register ve login tamamlandı twit işlevleri kontrol ediliyor

+++ src/main/java/com/workintech/twitter/service/CommentServiceImpl.java
@@ -3,15 +3,23 @@ package com.workintech.twitter.service;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.workintech.twitter.dto.request.CommentPatchRequestDto;
 import com.workintech.twitter.dto.request.CommentRequestDto;
 import com.workintech.twitter.dto.response.CommentResponseDto;
 import com.workintech.twitter.entity.Comment;
+import com.workintech.twitter.entity.Tweet;
+import com.workintech.twitter.entity.User;
 import com.workintech.twitter.exceptions.CommentNotFoundException;
+import com.workintech.twitter.exceptions.TweetNotFoundException;
+import com.workintech.twitter.exceptions.UserNotFoundException;
 import com.workintech.twitter.mapper.CommentMapper;
 import com.workintech.twitter.repository.CommentRepository;
+import com.workintech.twitter.repository.TweetRepository;
+import com.workintech.twitter.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -21,6 +29,12 @@ public class CommentServiceImpl implements CommentService {
     
     @Autowired
     private final CommentRepository commentRepository;
+
+    @Autowired
+    private final UserRepository userRepository;
+
+    @Autowired
+    private final TweetRepository tweetRepository;
 
     @Autowired
     private final CommentMapper commentMapper;
@@ -46,9 +60,21 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public CommentResponseDto create(CommentRequestDto commentRequestDto) {
-       Comment comment = commentMapper.toEntity(commentRequestDto);
-       comment = commentRepository.save(comment);
-       return commentMapper.toCommentResponseDto(comment);         
+      Comment comment = commentMapper.toEntity(commentRequestDto);
+
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    String email = authentication.getName();
+    User user = userRepository.findByEmail(email)
+        .orElseThrow(() -> new UserNotFoundException("Kullanıcı bulunamadı"));
+
+    Tweet tweet = tweetRepository.findById(commentRequestDto.tweetId())
+        .orElseThrow(() -> new TweetNotFoundException("Tweet bulunamadı"));
+
+    comment.setUser(user);
+    comment.setTweet(tweet);
+
+    comment = commentRepository.save(comment);
+    return commentMapper.toCommentResponseDto(comment);      
     }
 
     @Override
@@ -57,6 +83,8 @@ public class CommentServiceImpl implements CommentService {
         Optional<Comment> optComment = commentRepository.findById(id);
         if(optComment.isPresent()) {
             comment.setId(id);
+            comment.setUser(optComment.get().getUser());
+            comment.setTweet(optComment.get().getTweet());
             commentRepository.save(comment);
             return commentMapper.toCommentResponseDto(comment);
         }
@@ -69,7 +97,7 @@ public class CommentServiceImpl implements CommentService {
     
          Comment comentToUpdate = commentRepository
         .findById(id)
-        .orElseThrow(()-> new CommentNotFoundException(id+" id'li comment bulunamadı."));
+        .orElseThrow(()-> new CommentNotFoundException(id+" id'li comment bulunamadı."));      
 
         commentMapper.updateEntity(comentToUpdate, commentPatchRequestDto);
         commentRepository.save(comentToUpdate);
@@ -79,7 +107,18 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public void deleteById(Long id) {
-      commentRepository.deleteById(id);
+       Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    String email = authentication.getName();
+    User user = userRepository.findByEmail(email)
+        .orElseThrow(() -> new UserNotFoundException("Kullanıcı bulunamadı"));
+
+    Comment comment = commentRepository.findById(id)
+        .orElseThrow(() -> new CommentNotFoundException(id + " id'li comment bulunamadı."));
+if (comment.getUser() == null || user.getId() != comment.getUser().getId()) {
+    throw new RuntimeException("Bu yorumu silmeye yetkiniz yok.");
+}
+
+    commentRepository.deleteById(id);
     }
 
 }
