+++ src/main/java/com/workintech/twitter/service/TweetServiceImpl.java
@@ -4,15 +4,24 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import com.workintech.twitter.dto.request.TweetPatchRequestDto;
 import com.workintech.twitter.dto.request.TweetRequestDto;
 import com.workintech.twitter.dto.response.TweetResponseDto;
+import com.workintech.twitter.dto.response.UserResponseDto;
 import com.workintech.twitter.entity.Tweet;
+import com.workintech.twitter.entity.User;
 import com.workintech.twitter.exceptions.TweetNotFoundException;
+import com.workintech.twitter.exceptions.UnauthorizedDeleteException;
+import com.workintech.twitter.exceptions.UserNotFoundException;
 import com.workintech.twitter.mapper.TweetMapper;
+import com.workintech.twitter.mapper.UserMapper;
 import com.workintech.twitter.repository.TweetRepository;
+import com.workintech.twitter.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
 
@@ -24,7 +33,13 @@ public class TweetServiceImpl implements TweetService {
     private final TweetRepository tweetRepository;
 
     @Autowired
+    UserRepository userRepository;
+
+    @Autowired
     private final TweetMapper tweetMapper;
+
+    @Autowired
+    private final UserMapper userMapper;
 
     @Override
     public List<TweetResponseDto> getAll() {
@@ -48,21 +63,30 @@ public class TweetServiceImpl implements TweetService {
     @Override
     public TweetResponseDto create(TweetRequestDto tweetRequestDto) {
        Tweet tweet = tweetMapper.toEntity(tweetRequestDto);
+
+          Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    String email = authentication.getName();
+
+ User user = userRepository.findByEmail(email).orElseThrow(() -> new UserNotFoundException("Kullanıcı bulunamadı"));
+
+ tweet.setUser(user);
+ 
        tweet = tweetRepository.save(tweet);
        return tweetMapper.toTweetResponseDto(tweet); 
     }
 
     @Override
     public TweetResponseDto replaceOrCreate(Long id, TweetRequestDto tweetRequestDto) {
-        Tweet tweet = tweetMapper.toEntity(tweetRequestDto);
-        Optional<Tweet> optTweet = tweetRepository.findById(id);
-        if(optTweet.isPresent()) {
-            tweet.setId(id);
-            tweetRepository.save(tweet);
-            return tweetMapper.toTweetResponseDto(tweet);
-        }
+          Tweet tweet = tweetMapper.toEntity(tweetRequestDto);
+    Optional<Tweet> optTweet = tweetRepository.findById(id);
+    if(optTweet.isPresent()) {
+        tweet.setId(id);
+        tweet.setUser(optTweet.get().getUser());
         tweetRepository.save(tweet);
         return tweetMapper.toTweetResponseDto(tweet);
+    }
+    tweetRepository.save(tweet);
+    return tweetMapper.toTweetResponseDto(tweet);
     }
 
     @Override
@@ -78,7 +102,35 @@ public class TweetServiceImpl implements TweetService {
 
     @Override
     public void deleteById(Long id) {
+
+         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    String email = authentication.getName();
+
+User user = userRepository.findByEmail(email).orElseThrow(() -> new UserNotFoundException("Kullanıcı bulunamadı"));
+
+  Tweet tweet = tweetRepository.findById(id)
+        .orElseThrow(() -> new TweetNotFoundException(id + " id'li tweet bulunamadı."));
+
+    if (tweet.getUser() == null || user.getId() != tweet.getUser().getId()) {
+        throw new UnauthorizedDeleteException("Bu tweet'i silmeye yetkiniz yok.",HttpStatus.FORBIDDEN );
+    }
+
     tweetRepository.deleteById(id);
+    
+    }
+    
+    @Override
+    public List<UserResponseDto> getLikedUsers(Long id) {
+       Optional<Tweet> optTweet = tweetRepository.findById(id);
+       if(optTweet.isPresent()) {
+        return optTweet
+        .get()
+        .getLikes()        
+        .stream()
+        .map(like -> userMapper.toUserResponseDto(like.getUser()))
+        .toList();
+       }
+       throw new TweetNotFoundException(id + " id'li tweet bulunamadı.");
     }
 
 }
